Refactor effect group formatting helpers
